# frontend/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "D-WISE Tool Suite Backend API",
    "description": "The REST API for the D-WISE Tool Suite Backend.",
    "version": "beta_mwp_1"
  },
  "paths": {
...
    "/user/me": {
      "get": {
        "tags": ["user"],
        "summary": "Returns the current (logged in) user",
        "operationId": "get_me",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserRead" } } }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/user/{user_id}": {
      "get": {
        "tags": ["user"],
        "summary": "Returns the User with the given ID if it exists",
        "operationId": "get_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "User Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PublicUserRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "patch": {
        "tags": ["user"],
        "summary": "Updates the User with the given ID if it exists",
        "operationId": "update_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "User Id" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserUpdate" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "delete": {
        "tags": ["user"],
        "summary": "Removes the User with the given ID if it exists",
        "operationId": "delete_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "User Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": ["user"],
        "summary": "Returns all Users that exist in the system",
        "operationId": "get_all",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer", "maximum": 10000000, "minimum": 0 }, { "type": "null" }],
              "title": "Skip",
              "description": "The number of elements to skip (offset)"
            },
            "description": "The number of elements to skip (offset)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer", "maximum": 1000, "minimum": 1 }, { "type": "null" }],
              "title": "Limit",
              "description": "The maximum number of returned elements"
            },
            "description": "The maximum number of returned elements"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/PublicUserRead" },
                  "title": "Response User-Get All"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/user/{user_id}/project": {
      "get": {
        "tags": ["user"],
        "summary": "Returns all Projects of the User with the given ID",
        "operationId": "get_user_projects",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "User Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ProjectRead" },
                  "title": "Response User-Get User Projects"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/user/{user_id}/recent_activity": {
      "get": {
        "tags": ["user"],
        "summary": "Returns the top k sdoc ids that the User recently modified (annotated)",
        "operationId": "recent_activity",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "User Id" } },
          { "name": "k", "in": "query", "required": true, "schema": { "type": "integer", "title": "K" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/AnnotationDocumentRead" },
                  "title": "Response User-Recent Activity"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/project/{proj_id}/user/{user_id}": {
      "patch": {
        "tags": ["project"],
        "summary": "Associates an existing User to the Project with the given ID if it exists",
        "operationId": "associate_user_to_project",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "proj_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Proj Id" } },
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "User Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "delete": {
        "tags": ["project"],
        "summary": "Dissociates the Users with the Project with the given ID if it exists",
        "operationId": "dissociate_user_from_project",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "proj_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Proj Id" } },
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "User Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/project/{proj_id}/user": {
      "get": {
        "tags": ["project"],
        "summary": "Returns all Users of the Project with the given ID",
        "operationId": "get_project_users",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "proj_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Proj Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/UserRead" },
                  "title": "Response Project-Get Project Users"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/sdoc/{sdoc_id}": {
      "get": {
        "tags": ["sourceDocument"],
        "summary": "Returns the SourceDocument with the given ID if it exists",
        "operationId": "get_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "sdoc_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Sdoc Id" } },
          {
            "name": "only_if_finished",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean", "default": true, "title": "Only If Finished" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/SourceDocumentWithDataRead" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "delete": {
        "tags": ["sourceDocument"],
        "summary": "Removes the SourceDocument with the given ID if it exists",
        "operationId": "delete_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "sdoc_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Sdoc Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SourceDocumentRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "patch": {
        "tags": ["sourceDocument"],
        "summary": "Updates the SourceDocument with the given ID.",
        "operationId": "update_sdoc",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "sdoc_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Sdoc Id" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SourceDocumentUpdate" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SourceDocumentRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/doctag/{tag_id}": {
      "get": {
        "tags": ["documentTag"],
        "summary": "Returns the DocumentTag with the given ID.",
        "operationId": "get_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "tag_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Tag Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentTagRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "patch": {
        "tags": ["documentTag"],
        "summary": "Updates the DocumentTag with the given ID.",
        "operationId": "update_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "tag_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Tag Id" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentTagUpdate" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentTagRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "delete": {
        "tags": ["documentTag"],
        "summary": "Deletes the DocumentTag with the given ID.",
        "operationId": "delete_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "tag_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Tag Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentTagRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/span/{span_id}": {
      "get": {
        "tags": ["spanAnnotation"],
        "summary": "Returns the SpanAnnotation with the given ID.",
        "operationId": "get_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "span_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Span Id" } },
          {
            "name": "resolve",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Resolve Code",
              "description": "If true, the current_code_id of the SpanAnnotation gets resolved and replaced by the respective Code entity",
              "default": true
            },
            "description": "If true, the current_code_id of the SpanAnnotation gets resolved and replaced by the respective Code entity"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/SpanAnnotationRead" },
                    { "$ref": "#/components/schemas/SpanAnnotationReadResolved" }
                  ],
                  "title": "Response Spanannotation-Get By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "patch": {
        "tags": ["spanAnnotation"],
        "summary": "Updates the SpanAnnotation with the given ID.",
        "operationId": "update_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "span_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Span Id" } },
          {
            "name": "resolve",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Resolve Code",
              "description": "If true, the current_code_id of the SpanAnnotation gets resolved and replaced by the respective Code entity",
              "default": true
            },
            "description": "If true, the current_code_id of the SpanAnnotation gets resolved and replaced by the respective Code entity"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SpanAnnotationUpdateWithCodeId" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/SpanAnnotationRead" },
                    { "$ref": "#/components/schemas/SpanAnnotationReadResolved" }
                  ],
                  "title": "Response Spanannotation-Update By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "delete": {
        "tags": ["spanAnnotation"],
        "summary": "Deletes the SpanAnnotation with the given ID.",
        "operationId": "delete_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "span_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Span Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/SpanAnnotationRead" },
                    { "$ref": "#/components/schemas/SpanAnnotationReadResolved" }
                  ],
                  "title": "Response Spanannotation-Delete By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/span/code/{code_id}/user/{user_id}": {
      "get": {
        "tags": ["spanAnnotation"],
        "summary": "Returns SpanAnnotations with the given Code of the User with the given ID",
        "operationId": "get_by_user_code",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "code_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Code Id" } },
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "User Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/SpanAnnotationReadResolved" },
                  "title": "Response Spanannotation-Get By User Code"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/spangroup/{span_group_id}": {
      "get": {
        "tags": ["spanGroup"],
        "summary": "Returns the SpanGroup with the given ID.",
        "operationId": "get_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "span_group_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Span Group Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [{ "$ref": "#/components/schemas/SpanGroupRead" }, { "type": "null" }],
                  "title": "Response Spangroup-Get By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "patch": {
        "tags": ["spanGroup"],
        "summary": "Updates the SpanGroup with the given ID.",
        "operationId": "update_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "span_group_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Span Group Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SpanGroupUpdate" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [{ "$ref": "#/components/schemas/SpanGroupRead" }, { "type": "null" }],
                  "title": "Response Spangroup-Update By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "delete": {
        "tags": ["spanGroup"],
        "summary": "Deletes the SpanGroup with the given ID.",
        "operationId": "delete_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "span_group_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Span Group Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [{ "$ref": "#/components/schemas/SpanGroupRead" }, { "type": "null" }],
                  "title": "Response Spangroup-Delete By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/bbox/{bbox_id}": {
      "get": {
        "tags": ["bboxAnnotation"],
        "summary": "Returns the BBoxAnnotation with the given ID.",
        "operationId": "get_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "bbox_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Bbox Id" } },
          {
            "name": "resolve",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Resolve Code",
              "description": "If true, the current_code_id of the SpanAnnotation gets resolved and replaced by the respective Code entity",
              "default": true
            },
            "description": "If true, the current_code_id of the SpanAnnotation gets resolved and replaced by the respective Code entity"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/BBoxAnnotationRead" },
                    { "$ref": "#/components/schemas/BBoxAnnotationReadResolvedCode" }
                  ],
                  "title": "Response Bboxannotation-Get By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "patch": {
        "tags": ["bboxAnnotation"],
        "summary": "Updates the BBoxAnnotation with the given ID.",
        "operationId": "update_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "bbox_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Bbox Id" } },
          {
            "name": "resolve",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Resolve Code",
              "description": "If true, the current_code_id of the SpanAnnotation gets resolved and replaced by the respective Code entity",
              "default": true
            },
            "description": "If true, the current_code_id of the SpanAnnotation gets resolved and replaced by the respective Code entity"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/BBoxAnnotationUpdateWithCodeId" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/BBoxAnnotationRead" },
                    { "$ref": "#/components/schemas/BBoxAnnotationReadResolvedCode" }
                  ],
                  "title": "Response Bboxannotation-Update By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "delete": {
        "tags": ["bboxAnnotation"],
        "summary": "Deletes the BBoxAnnotation with the given ID.",
        "operationId": "delete_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "bbox_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Bbox Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/BBoxAnnotationRead" },
                    { "$ref": "#/components/schemas/BBoxAnnotationReadResolvedCode" }
                  ],
                  "title": "Response Bboxannotation-Delete By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/bbox/code/{code_id}/user/{user_id}": {
      "get": {
        "tags": ["bboxAnnotation"],
        "summary": "Returns BBoxAnnotations with the given Code of the User with the given ID",
        "operationId": "get_by_user_code",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "code_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Code Id" } },
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "User Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/BBoxAnnotationRead" },
                  "title": "Response Bboxannotation-Get By User Code"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": { "items": { "$ref": "#/components/schemas/ValidationError" }, "type": "array", "title": "Detail" }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserRead": {
        "properties": {
          "email": { "type": "string", "format": "email", "title": "Email", "description": "E-Mail of the User" },
          "first_name": { "type": "string", "title": "First Name", "description": "First name of the User" },
          "last_name": { "type": "string", "title": "Last Name", "description": "Last name of the User" },
          "id": { "type": "integer", "title": "Id", "description": "ID of the User" },
          "created": {
            "type": "string",
            "format": "date-time",
            "title": "Created",
            "description": "Created timestamp of the User"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "title": "Updated",
            "description": "Updated timestamp of the User"
          }
        },
        "type": "object",
        "required": ["email", "first_name", "last_name", "id", "created", "updated"],
        "title": "UserRead"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "/authentication/login" } }
      }
    }
  }
}